+++ CHANGELOG.md
@@ -6,10 +6,12 @@ All notable changes to this project will be documented in this file.
 ## [0.0.1]
 ### 2018-03-10
 ### Added
-* Added timepicker.js components -CL
+* Added currentTask.js Component -CL
+* Added timepicker.js component -CL
 * Added Changelog -CL
 * Added Readme -CL
 
 ### Updated
+* Updated App.js (Display currentTask) -CL
 * Updated App.js (Display timepicker) -CL
 * Change origin readme to ReactNativeNotes -CL